+++ source/CHANGELOG.md
@@ -10,7 +10,7 @@ NEW:
 * Added support for importing SWMM time series file format 
 * Wave project file (WVP):
   * Added support for additional series options #117
-  * When saving to WVP, changed series titles are now also saved
+  * When saving to WVP, series options are now also saved
 
 FIXED:
 * Analysis function Accumulate now repects the input interpretation